# packages/api-proxy/swagger/swagger-spec-v3-klaytn.yaml
openapi: 3.0.0
info:
  title: KLAYTN
  description: Henesis API Proxy - v3 - klaytn OpenAPI Documentation
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  schemas:
    InvalidAccessIpException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: ip '106.101.128.133' is not allowed
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4008
      required:
        - message
        - code
    InvalidAccessTokenException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: expired or invalid access token and signature
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4001
      required:
        - message
        - code
    AccessTokenNotProvidedException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: access token not provided
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4001
      required:
        - message
        - code
...
    PaginationMetadata:
      type: object
      properties:
        nextUrl:
          type: string
          description: 다음 pagination url
          example: http://localhost:3000/api/v3/ethereum/transfers?page=2
        previousUrl:
          type: string
          description: 이전 pagination url
          example: http://localhost:3000/api/v3/ethereum/transfers?page=1
        totalCount:
          type: number
          description: 총 갯수
          example: 5
      required:
        - nextUrl
        - previousUrl
        - totalCount
...
    TransactionIdNotFoundException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: transaction id '1' does not exists
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4000
      required:
        - message
        - code
    TransactionDTO:
      type: object
      properties:
        id:
          type: string
          description: >-
            트랜잭션 ID (Henesis Wallet에서 부여하는 트랜잭션의 고유 ID입니다, 블록체인에서 부여하는 트랜잭션 해시와
            다른 개념입니다)
          example: b549bfaaa74d1c4244ecc655738b1984
        blockchain:
          enum:
            - ETHEREUM
            - KLAYTN
            - BITCOIN
            - LITECOIN
            - FILECOIN
            - BINANCE_SMART_CHAIN
            - BITCOIN_CASH
          type: string
          description: 메인넷 종류
          example: ETHEREUM
        hash:
          type: string
          description: 트랜잭션 해시 (트랜잭션 상태가 REQUESTED일 때는 존재하지 않습니다)
          example: '0x8a1ef722a2884d53a86874c8ba19bcf267bb0b8a81da00c14ddc6457cf1eaf96'
        error:
          type: string
          description: 트랜잭션 전송 시 발생한 에러
          example: 'null'
        status:
          enum:
            - PENDING_APPROVAL
            - REJECTED
            - REQUESTED
            - PENDING
            - FAILED
            - MINED
            - REVERTED
            - INTERNAL_REVERTED
            - CONFIRMED
            - REPLACED
          type: string
          description: 트랜잭션 상태
          example: CONFIRMED
        fee:
          type: string
          description: 트랜잭션 수수료
          example: '10000000'
        hopAddress:
          type: string
          description: 홉 주소
          example: '0x1AA2705a26452cC22430F31A5c85974bBEDDe5a5'
        createdAt:
          type: string
          description: '트랜잭션 생성 시간 (형식: ms, UNIX time)'
          example: '1614582928222'
        updatedAt:
          type: string
          description: '트랜잭션 상태가 변한 시간 (형식: ms, UNIX time)'
          example: '1612411724023'
      required:
        - id
        - blockchain
        - status
        - createdAt
        - updatedAt
    GetManyResponseForTransactionDTO:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDTO'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
      required:
        - results
    WalletNotFoundException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: wallet 'c27732256a1f8d45df5aec31ba7ef85cd' does not exist
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4000
      required:
        - message
        - code
    NoWalletNameException:
      type: object
      properties:
        message:
          type: string
          description: API 에러 메시지
          example: wallet name should not be empty
        code:
          type: number
          description: 헤네시스 API 에러 코드
          example: 4000
      required:
        - message
        - code
...
    WalletDTO:
      type: object
      properties:
        id:
          type: string
          description: 지갑 ID
          example: ae40b1b3dd953e5592c21e58be30d807
        name:
          type: string
          description: 지갑 이름
          example: ethereum-wallet
        address:
          type: string
          description: 지갑 주소
          example: '0xdde12e85d7f5d0003eff13f7a3cf6068b5a41454'
        blockchain:
          enum: &ref_0
            - ETHEREUM
            - KLAYTN
            - BITCOIN
            - LITECOIN
            - FILECOIN
            - BINANCE_SMART_CHAIN
            - BITCOIN_CASH
          type: string
          description: 메인넷 종류
          example: ETHEREUM
        createdAt:
          type: string
          description: '지갑 생성 시간 (형식: ms, UNIX time)'
          example: '1612411568760'
        updatedAt:
          type: string
          description: '지갑 정보 수정 시간 (단위: ms, UNIX time)'
          example: '1612411724023'
        status:
          enum: &ref_1
            - ACTIVE
            - CREATING
            - FAILED
            - INACTIVE
          type: string
          description: 지갑 상태
          example: ACTIVE
        whitelistActivated:
          type: boolean
          description: 출금 주소 화이트리스팅의 활성화 유무
          example: false
        version:
          type: string
          description: 컨트랙트 버전
          example: v4
      required:
        - id
        - name
        - address
        - blockchain
        - createdAt
        - updatedAt
        - status
    BalanceDTO:
      type: object
      properties:
        coinId:
          type: number
          description: Henesis에서 부여한 Coin의 ID
          example: 11
        amount:
          type: string
          description: '확정된 잔액 (단위: wei, peb)'
          example: '1000000000000'
        spendableAmount:
          type: string
          description: '출금 가능한 잔액 (= 총 잔액 - 확정되지 않은 출금 요청액)(단위: wei, peb, jager)'
          example: '1000000000000'
        name:
          type: string
          description: 코인 이름
          example: ethereum
        ticker:
          type: string
          description: 코인의 기호
          example: ETH
        decimals:
          type: number
          description: 코인의 소수점 자릿수
          example: 18
      required:
        - coinId
        - amount
        - spendableAmount
        - name
        - ticker
        - decimals
...
    KeyDTO:
      type: object
      properties:
        address:
          type: string
          description: ''
          example: '0x4ef3ba60c8710f45371835cddafabf33daa83e1d'
        pub:
          type: string
          description: pub
          example: >-
            0x31bd93d049fefed19b640c8069046c223126505754b9a57f5df43a89b104d92c8d4be4f51a6b5bb08a3ec6c2ff022e8ff018bad52ee05fa81b4eeae16a0e2db1
        feeDelegationEnabled:
          type: boolean
          description: feeDelegationEnabled
          example: false
        keyFile:
          type: string
          description: keyFile
          example: '1'
        keyId:
          type: string
          description: keyId
          example: 52e779750bb1330d2f23439c6da821ee
      required:
        - pub
    MasterWalletDTO:
      type: object
      properties:
        id:
          type: string
          description: 지갑 ID
          example: ae40b1b3dd953e5592c21e58be30d807
        name:
          type: string
          description: 지갑 이름
          example: ethereum-wallet
        address:
          type: string
          description: 지갑 주소
          example: '0xdde12e85d7f5d0003eff13f7a3cf6068b5a41454'
        blockchain:
          enum: *ref_0
          type: string
          description: 메인넷 종류
          example: ETHEREUM
        createdAt:
          type: string
          description: '지갑 생성 시간 (형식: ms, UNIX time)'
          example: '1612411568760'
        updatedAt:
          type: string
          description: '지갑 정보 수정 시간 (단위: ms, UNIX time)'
          example: '1612411724023'
        status:
          enum: *ref_1
          type: string
          description: 지갑 상태
          example: ACTIVE
        whitelistActivated:
          type: boolean
          description: 출금 주소 화이트리스팅의 활성화 유무
          example: false
        version:
          type: string
          description: 컨트랙트 버전
          example: v4
        accountKey:
          description: KEY
          example: &ref_19
            address: '0x4ef3ba60c8710f45371835cddafabf33daa83e1d'
            pub: >-
              0x31bd93d049fefed19b640c8069046c223126505754b9a57f5df43a89b104d92c8d4be4f51a6b5bb08a3ec6c2ff022e8ff018bad52ee05fa81b4eeae16a0e2db1
            feeDelegationEnabled: false
            keyFile: '1'
            keyId: 52e779750bb1330d2f23439c6da821ee
          allOf:
            - $ref: '#/components/schemas/KeyDTO'
        encryptionKey:
          type: string
          description: 마스터 지갑 비밀번호를 복구하기 위해 암호화하는 데 쓰인 키
          example: f0880e48251dd2f8574a8c8a4d65311030185d5e451e8a474b89c2b473b5d953
      required:
        - id
        - name
        - address
        - blockchain
        - createdAt
        - updatedAt
        - status
        - accountKey
        - encryptionKey
    ChangeWalletNameRequestDTO:
      type: object
      properties:
        name:
          type: string
          description: 지갑의 새로운 이름
          example: name
      required:
        - name
    SendCoinRequestDTO:
      type: object
      properties:
        ticker:
          type: string
          description: 암호화폐의 기호 (ticker)
          example: ETH
        to:
          type: string
          description: 암호화폐를 받을 지갑 주소
          example: '0xab28d146e860e0b132695c941f706d783a158345'
        amount:
          type: string
          description: '암호화폐의 양 (단위: wei, peb)'
          example: '100000000'
        passphrase:
          type: string
          description: 지갑의 비밀번호
          example: passphrase
        gasPrice:
          type: string
          description: >-
            트랜잭션에 사용할 gas price (단위: wei) null일 경우, Henesis가 자동으로 5분 안에 채굴될 수 있는
            값으로 설정합니다.
          example: '8000000000'
        gasLimit:
          type: string
          description: 트랜잭션에 사용할 gas limit. null일 경우, Henesis가 자동으로 설정합니다.
          example: '500000'
        isHopTransaction:
          type: boolean
          description: 컨트랙트를 통한 입금을 원치 않는 경우, 컨트랙트가 아닌 EOA로 출금을 진행 합니다.
          example: false
        metadata:
          type: string
          description: 기타 정보 기록용 메타 데이터 (255자 제한)
          example: metadata
      required:
        - ticker
        - to
        - amount
        - passphrase
...
paths:
...
  /api/v3/klaytn/transactions/{transactionId}:
    get:
      operationId: klaytn-TransactionsController_getTransaction
      summary: 개별 트랜잭션 조회하기
      description: 내가 발생시킨 특정 트랜잭션의 정보를 조회합니다.
      parameters:
        - &ref_13
          name: X-Henesis-Secret
          in: header
          description: 대시보드를 통해 발급 받은 API Secret을 입력합니다.
          schema:
            type: string
        - &ref_14
          name: Authorization
          in: header
          description: 대시보드를 통해 발급 받은 Access Token을 입력합니다. e.g. Bearer <TOKEN\>
          schema:
            example: Bearer <TOKEN>
            default: 'Bearer '
            type: string
        - name: transactionId
          required: true
          in: path
          description: >-
            트랜잭션 ID (Henesis Wallet에서 부여하는 트랜잭션의 고유 ID입니다. 온체인상 트랜잭션 해시와 다른
            개념입니다.)
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
              example: &ref_15
                id: b549bfaaa74d1c4244ecc655738b1984
                blockchain: ETHEREUM
                hash: >-
                  0x8a1ef722a2884d53a86874c8ba19bcf267bb0b8a81da00c14ddc6457cf1eaf96
                error: 'null'
                status: CONFIRMED
                fee: '10000000'
                hopAddress: '0x1AA2705a26452cC22430F31A5c85974bBEDDe5a5'
                createdAt: '1614582928222'
                updatedAt: '1612411724023'
          description: ''
        '400':
          description: transaction id가 없을 때 발생합니다
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionIdNotFoundException'
              example:
                message: transaction id '1' does not exists
                code: 4000
        '401': &ref_16
          description: 인증이 실패했을 때 발생하는 에러입니다.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/InvalidAccessTokenException'
                  - $ref: '#/components/schemas/AccessTokenNotProvidedException'
                  - $ref: '#/components/schemas/InvalidAccessIpException'
              examples:
                InvalidAccessTokenException:
                  value: *ref_8
                AccessTokenNotProvidedException:
                  value: *ref_9
                InvalidAccessIpException:
                  value: *ref_10
      tags:
        - transactions
  /api/v3/klaytn/transactions:
    get:
      operationId: klaytn-TransactionsController_getTransactions
      x-readme:
        explorer-enabled: false
        samples-languages:
          - curl
          - node
          - java
          - python
          - go
      summary: 트랜잭션 목록 조회하기
      description: 트랜잭션 목록을 조회합니다.
      parameters:
        - *ref_13
        - *ref_14
        - name: address
          required: false
          in: query
          description: 트랜잭션 주소
          schema:
            type: string
        - name: toAddress
          required: false
          in: query
          description: 받는 주소
          schema:
            type: string
        - name: fromAddress
          required: false
          in: query
          description: 보내는 주소
          schema:
            type: string
        - name: transactionHash
          required: false
          in: query
          description: 트랜잭션 해시
          schema:
            type: string
        - name: start
          required: false
          in: query
          description: 해당 시점 이후의 트랜잭션 조회
          schema:
            type: string
        - name: end
          required: false
          in: query
          description: 해당 시점 이전의 트랜잭션 조회
          schema:
            type: string
        - name: status
          required: false
          in: query
          description: 트랜잭션 상태
          schema:
            type: string
        - name: statuses
          required: false
          in: query
          description: 트랜잭션 상태
          schema:
            type: array
            items:
              type: string
        - name: types
          required: false
          in: query
          description: 트랜잭션 종류
          schema:
            type: array
            items:
              type: string
        - name: keyId
          required: false
          in: query
          description: 키 ID
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetManyResponseForTransactionDTO'
              example:
                pagination: *ref_11
                results:
                  - *ref_15
          description: ''
        '401': *ref_16
      tags:
        - transactions
  /api/v3/klaytn/wallets:
    get:
      operationId: klaytn-WalletsController_getMasterWallets
      x-readme:
        explorer-enabled: false
        samples-languages:
          - curl
          - node
          - java
          - python
          - go
      summary: 전체 마스터 지갑 목록 조회하기
      description: 모든 지갑의 목록을 조회합니다.
      parameters:
        - &ref_17
          name: X-Henesis-Secret
          in: header
          description: 대시보드를 통해 발급 받은 API Secret을 입력합니다.
          schema:
            type: string
        - &ref_18
          name: Authorization
          in: header
          description: 대시보드를 통해 발급 받은 Access Token을 입력합니다. e.g. Bearer <TOKEN\>
          schema:
            example: Bearer <TOKEN>
            default: 'Bearer '
            type: string
        - name: name
          required: false
          in: query
          description: 지갑 이름
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MasterWalletDTO'
          description: ''
        '401': &ref_20
          description: 인증이 실패했을 때 발생하는 에러입니다.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/InvalidAccessTokenException'
                  - $ref: '#/components/schemas/AccessTokenNotProvidedException'
                  - $ref: '#/components/schemas/InvalidAccessIpException'
              examples:
                InvalidAccessTokenException:
                  value: *ref_8
                AccessTokenNotProvidedException:
                  value: *ref_9
                InvalidAccessIpException:
                  value: *ref_10
      tags:
        - wallets
  /api/v3/klaytn/wallets/{masterWalletId}:
    get:
      operationId: klaytn-WalletsController_getMasterWallet
      x-readme:
        explorer-enabled: false
        samples-languages:
          - curl
          - node
          - java
          - python
          - go
      summary: 마스터 지갑 정보 조회하기
      description: 특정 마스터 지갑의 정보를 조회합니다.
      parameters:
        - *ref_17
        - *ref_18
        - name: masterWalletId
          required: true
          in: path
          description: 마스터 지갑 ID
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletDTO'
              example: &ref_22
                id: ae40b1b3dd953e5592c21e58be30d807
                name: ethereum-wallet
                address: '0xdde12e85d7f5d0003eff13f7a3cf6068b5a41454'
                blockchain: ETHEREUM
                createdAt: '1612411568760'
                updatedAt: '1612411724023'
                status: ACTIVE
                whitelistActivated: false
                version: v4
                accountKey: *ref_19
                encryptionKey: >-
                  f0880e48251dd2f8574a8c8a4d65311030185d5e451e8a474b89c2b473b5d953
          description: ''
        '400':
          description: 해당하는 id의 지갑이 없을 때 발생합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletNotFoundException'
              example: &ref_21
                message: wallet 'c27732256a1f8d45df5aec31ba7ef85cd' does not exist
                code: 4000
        '401': *ref_20
      tags:
        - wallets
  /api/v3/klaytn/wallets/{masterWalletId}/balance:
    get:
      operationId: klaytn-WalletsController_getMasterWalletBalance
      x-readme:
        explorer-enabled: false
        samples-languages:
          - curl
          - node
          - java
          - python
          - go
      summary: 마스터 지갑 잔고 조회하기
      description: 특정 마스터 지갑의 잔고를 조회합니다.
      parameters:
        - *ref_17
        - *ref_18
        - name: masterWalletId
          required: true
          in: path
          description: 마스터 지갑 ID
          schema:
            type: string
        - name: ticker
          required: false
          in: query
          description: 코인의 기호
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BalanceDTO'
          description: ''
        '400':
          description: 해당하는 id의 지갑이 없을 때 발생합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletNotFoundException'
              example: *ref_21
        '401': *ref_20
      tags:
        - wallets
  /api/v3/klaytn/wallets/{masterWalletId}/name:
    patch:
      operationId: klaytn-WalletsController_changeMasterWalletName
      x-readme:
        explorer-enabled: false
        samples-languages:
          - curl
          - node
          - java
          - python
          - go
      summary: 마스터 지갑 이름 변경하기
      description: 특정 마스터 지갑의 이름을 변경합니다.
      parameters:
        - *ref_17
        - *ref_18
        - name: masterWalletId
          required: true
          in: path
          description: 마스터 지갑 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeWalletNameRequestDTO'
      responses:
        '200':
          description: ''
        '204':
          description: ''
        '400':
          description: 다음과 같은 bad request 에러가 발생할 수 있습니다.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/WalletNotFoundException'
                  - $ref: '#/components/schemas/NoWalletNameException'
              examples:
                WalletNotFoundException:
                  value: *ref_21
                NoWalletNameException:
                  value: &ref_23
                    message: wallet name should not be empty
                    code: 4000
        '401': *ref_20
      tags:
        - wallets
  /api/v3/klaytn/wallets/{masterWalletId}/transfer:
    post:
      operationId: klaytn-WalletsController_sendMasterWalletCoin
      x-readme:
        explorer-enabled: false
        samples-languages:
          - curl
          - node
          - java
          - python
          - go
      summary: 마스터 지갑에서 코인 전송하기
      description: 특정 마스터 지갑에서 가상자산을 송금합니다.
      parameters:
        - *ref_17
        - *ref_18
        - name: masterWalletId
          required: true
          in: path
          description: 마스터 지갑 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCoinRequestDTO'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
              example: *ref_15
          description: ''
        '400':
          description: 해당하는 id의 지갑이 없을 때 발생합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletNotFoundException'
              example: *ref_21
        '401': *ref_20
      tags:
        - wallets
...